requirements.txt
oauth2client==4.1.3
mysql-connector-python
python-dotenv
argparse
pandas
numpy
pygsheets
python-json-logger